+++ module-apps/application-settings-new/windows/KeypadLightWindow.hpp
@@ -4,11 +4,15 @@
 
 namespace gui
 {
+    class CheckBoxWithLabel;
 
     class KeypadLightWindow : public BaseSettingsWindow
     {
       public:
         KeypadLightWindow(app::Application *app);
         void buildInterface() override;
+
+      private:
+        std::vector<CheckBoxWithLabel *> boxes;
     };
 } // namespace gui