+++ changelog.md
@@ -7,6 +7,9 @@
 * `[gui]` Dynamic Windows building and handling implemented
 * `[audio]` Added A2DP playback backend
 * `[audio]` Added SendEvent audio API
+* `[calculator]` Add calculator application.
+* `[settings]` Create Messages window for apps and tools branch.
+
 
 ### Fixed